requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.13
psycopg2-binary==2.9.6
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pillow==9.5.0
opencv-python-headless==4.7.0.72
pyzbar==0.1.9
email-validator==2.0.0
gunicorn==20.1.0
alembic==1.10.4 